+++ src/api/app/views/layouts/webui/_personal_navigation.html.haml
@@ -16,15 +16,15 @@
     - else
       = link_to 'Create Home', new_project_path(name: User.current.home_project_name)
     |
-    \#{link_to 'Logout', user_logout_path, method: :post, id: 'logout-link'}
+    \#{link_to 'Logout', session_destroy_path, method: :delete, id: 'logout-link'}
   - elsif CONFIG['kerberos_mode']
-    = link_to 'Log In', user_login_path
+    = link_to 'Log In', session_new_path
   - else
     - if CONFIG['proxy_auth_mode'] == :on
       - unless CONFIG['proxy_auth_register_page'].blank?
         = link_to 'Sign Up', "#{CONFIG['proxy_auth_register_page']}?%22"
         |
-      = link_to 'Log In', user_login_path, id: 'login-trigger'
+      = link_to 'Log In', session_new_path, id: 'login-trigger'
       #login-form
         = form_tag(CONFIG['proxy_auth_login_page'], method: :post, id: 'login_form', enctype: 'application/x-www-form-urlencoded') do
           %p
@@ -43,9 +43,9 @@
       - if can_register
         = link_to 'Sign Up', user_register_user_path
         |
-      = link_to 'Log In', user_login_path, id: 'login-trigger'
+      = link_to 'Log In', session_new_path, id: 'login-trigger'
       #login-form
-        = form_tag(user_do_login_path) do
+        = form_tag(session_create_path) do
           %p
             = label_tag(:username, 'Username')
             = text_field_tag(:username, '')
