+++ src/api/spec/routing/webui/users/requests_spec.rb
@@ -1,7 +1,8 @@
 require 'rails_helper'
 
-RSpec.describe '/users/:user/requests routes' do
+RSpec.describe '/users/:user/requests routes', type: :routing do
   let(:user) { Faker::Name.first_name }
+
   it do
     expect(get("/users/#{user}/requests"))
       .to route_to('webui/users/bs_requests#index', user: user)
