+++ src/api/spec/support/features/features_authentication.rb
@@ -1,6 +1,6 @@
 module FeaturesAuthentication
   def login(user, password = 'buildservice')
-    visit user_login_path
+    visit session_new_path
     expect(page).to have_text 'Please Log In'
     fill_in 'user-login', with: user.login
     fill_in 'user-password', with: password
@@ -10,7 +10,7 @@ module FeaturesAuthentication
   end
 
   def logout
-    visit user_logout_path
+    visit session_destroy_path
     expect(page).to have_no_link('link-to-user-home')
     User.current = nil
   end
