+++ src/api/test/test_helper.rb
@@ -207,7 +207,7 @@ module Webui
     def login_user(user, password, opts = {})
       # no idea why calling it twice would help
       WebMock.disable_net_connect!(allow_localhost: true)
-      visit user_login_path
+      visit session_new_path
       fill_in 'Username', with: user
       fill_in 'Password', with: password
       click_button 'Log In'
